mes-3649-restructuredFolderForWRTCPage
