+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { WaitingRoomToCarPage } from './waiting-room-to-car';
 import { EffectsModule } from '@ngrx/effects';
-import { WaitingRoomToCarAnalyticsEffects } from '../waiting-room-to-car/waiting-room-to-car.analytics.effects';
-import { ComponentsModule } from '../../components/common/common-components.module';
-import { WaitingRoomToCarComponentsModule } from './components/waiting-room-to-car.components.module';
+import { WaitingRoomToCarAnalyticsEffects } from '../waiting-room-to-car.analytics.effects';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { WaitingRoomToCarComponentsModule } from '../components/waiting-room-to-car.components.module';
 
 @NgModule({
   declarations: [
