+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.scss
